// client/src/components/Button/Button.tsx
import styled from 'styled-components';
export const Button = styled.button`
  display: block;
  margin: 2rem auto 0;
  padding: 10px 20px;
  background-color: #3c6e71;
  color: #fff;
  border: none;
  border-radius: 5px;
  cursor: pointer;
  font-size: 1rem;
  transition: background-color 0.3s;

  &:hover {
    background-color: #284b63;
  }
`;